Refactored top header bar into a Header component
Also, the current view is kept orange in the navbar inside of the
Header. This was achieved by using a higher-order-component withRouter
inside of App.js, which automatically injects props into the App
component including the history (ie info on which route we're on or have
gone to in the past). The current route is sent to the new Header component
as a prop which can use it to set a class on the link that matches the
current router which in turn colors that link text orange