Refactored styles to use rem instead of px and unified animation times
Now there is only on px value for the default font size, set in body css
and everything else that was previously a pixel width with use a rem
unit for width instead, where rem is proportional to the body css px
value (ie 1rem = 16px, 1.5rem = 24px, ect). This allows the whole page
to scale up or down easily by changing the font-size in body alone.

Instead of having animation times set in each css file, now index.css
has animation times set in one place for the whole page, making it easy
to adjust the animation timing of the whole page in one place.