Made page responsive
By changing the main pixel font-size (in html element css, last commit putting it it
body was a mistake), we can scale down the page, as well as adjust other
element widths to allow them to take up the full width in mobile view.

This is not a mobile-first design as the default styles are all desktop
styles, and then styles applied by the media call is the mobile style.
Mobile-first would be the opposite thing.