Proper two-way binding for form and child elements
Removed the use of context, and used React.Children.map to generate an
array of the children of Form and add set their value prop to Form state
as well as set their onChange handler to modify Form state.

Added a custom message prop for each input if it is invalid that pops up
below the input if it is invalid.