Fixed a few React warnings
It did not like errorMessage so renamed it errormessage
When passing rest of props to input element in Input, the invalid prop
being equal to a boolean was throwing a warning. This prop was used
already to set a class an actually did not need to be set as an
attribute on input, so used some fancy destructuring to get the rest of
the props left over after destructuring errormessage and invalid as
rest, and using that to set attribs for the ipnput.