Added baic search and filter bar functionality
On Browse tab, you now hav a bar that allows sorting and filtering by
whatever field is currently being sorte by. This is not styled or
cleaned up yet, and I had to move the state of MovieIndex wholly into
App for this to work, which broke the Feedback (it only shows once).